Adding local csv energy test data
Will be useful to have a csv file with real data for local testing
w/o the CMT data or the postgres db. Also created a conf file
to try using that data.

Since the real data has missing column values, I needed to change
the csv ingester to filter out bad rows by caching all the successfully
parsed rows in memory. This will hopefully not be an issue for any of
the small tests we are using csvparser for, and should be faster
to iterate through.

To test run ./bin/batch.sh conf/local_batch.yaml